requirements.txt
Django==5.0.6
django-htmx==1.17.3
django-parler==2.3
django-taggit==5.0.1
markdown==3.6
Pillow==10.4.0
